Move remove and update logic to RecipesList
