+++ client/src/components/recipes/RecipesList.svelte
@@ -1,6 +1,7 @@
 <script>
   import Card from '@/components/common/Card.svelte';
   import { createEventDispatcher } from 'svelte';
+  import { recipeApi } from '@/api';
 
   export let category = null;
   export let published = null;
@@ -21,6 +22,19 @@
     const isPublishedNotSelected = published === 0;
     return doPublishedMatch || isPublishedNotSelected;
   };
+
+  const remove = async id => {
+    if (!id) return alert('Delete failed, try again');
+    await recipeApi.remove(id);
+    dispatch('refetch');
+  };
+
+  const updateIsPublished = async (id, isPublished) => {
+    await recipeApi.updateIsPublished(id, isPublished);
+    const recipe = recipes.find(recipe => recipe.id === id);
+    recipe.isPublished = !recipe.isPublished;
+    recipes = [...recipes];
+  };
 </script>
 
 <section>
@@ -33,8 +47,8 @@
         time={preparationTime}
         {shouldDisplayBonusActions}
         {isPublished}
-        on:remove={() => dispatch('remove', { id })}
-        on:update={() => dispatch('update', { id, isPublished })}
+        on:remove={remove(id)}
+        on:update={updateIsPublished(id, isPublished)}
       />
     {/if}
   {/each}
